+++ src/main/java/io/github/mzmine/modules/tools/batchwizard/subparameters/BatchWizardHPLCParameters.java
@@ -23,7 +23,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.github.mzmine.modules.tools.batchwizard;
+package io.github.mzmine.modules.tools.batchwizard.subparameters;
 
 import com.google.common.collect.Range;
 import io.github.mzmine.modules.tools.batchwizard.WizardPreset.ChromatographyDefaults;
